backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
sqlparse==0.4.3
psycopg2-binary==2.9.1
gunicorn==20.1.0
